Clean up old Mesen C++ source, merge config, fix input shortcuts, save state in config
- Remove all legacy Mesen C++ source files (Core/, InteropDLL/, Lua/, EmbeddedRoms/)
- Merge RomConfig into Config as single .config file per game
- Fix gamepad button shortcut: PS Cross / Xbox A = Start on title/pause, normal during play
- Track pause state with rising edge detection for reliable shortcut toggling
- D-pad Up/Down always triggers Select for menu navigation
- Save state now stored inside config file instead of separate file
- Fullscreen mode with auto-detect controller type (PS vs Xbox)
- Loading screen with centered logo

+++ src/Core/PPU.cs
@@ -430,7 +430,7 @@ public class PPU {
         return NesPalette[paletteColorIndex % 64];
     }    
 
-    public void DrawFrame(int scale) {
+    public void DrawFrame(float scale) {
         for (int y = 0; y < ScreenHeight; y++) {
             for (int x = 0; x < ScreenWidth; x++) {
                 Color color = frameBuffer[y * ScreenWidth + x];
